+++ R/profile_analysis.R
@@ -43,22 +43,23 @@ profile_tables_ = function(items, a, b, design, domains, item_property)
 {
   design = design |> 
     inner_join(domains[,c('item_id',item_property)], by='item_id') |>
-    mutate(dcat = dense_rank(.data[[item_property]]))
+    mutate(dcat = dense_rank(.data[[item_property]])) |>
+    arrange(.data$booklet_id, .data$dcat, .data$item_id) 
   
-  dcat = distinct(design, .data[[item_property]], dcat )
+  domain_category_index = distinct(design, .data[[item_property]], dcat )
   
   design |>
     group_by(.data$booklet_id) |>
     do({
       prof_lmx = E_profile(b,a,.$first,.$last, split(1:nrow(.), .$dcat))
-      
+      dcat_bk = sort(unique(.$dcat))
       lapply(prof_lmx, function(mtx)
       {
-        tibble(dcat=1:ncol(mtx), expected_domain_score = mtx[1,] )}
+        tibble(dcat=dcat_bk, expected_domain_score = mtx[1,] )}
       ) |>
         bind_rows(.id='booklet_score') |>
         mutate(booklet_score = as.integer(.data$booklet_score)-1L) |>
-        inner_join(dcat, by='dcat')
+        inner_join(domain_category_index, by='dcat')
       
     }) |>
     ungroup() |>